src/1-collector-service/requirements.txt
pandas==1.3.5
numpy==1.21.4
pyyaml==6.0
requests==2.26.0
python-dotenv==0.19.2